+++ frontend/public/index.css
@@ -1,149 +1,168 @@
 @font-face {
-  font-family: "earthorbiter";
-  src: local("earthorbiter"),
-    url(/fonts/earthorbiteracad.ttf);
+	font-family: "earthorbiter";
+	src: local("earthorbiter"),
+		url(/fonts/earthorbiteracad.ttf);
 }
 
 @font-face {
-  font-family: "earthorbiter3";
-  src: local("earthorbiter3"),
-    url(/fonts/earthorbiter3dital.ttf);
+	font-family: "earthorbiter3";
+	src: local("earthorbiter3"),
+		url(/fonts/earthorbiter3dital.ttf);
 }
 
 @font-face {
-  font-family: "Countless";
-  src: local("Countless"),
-    url(/fonts/Countless.ttf);
+	font-family: "Countless";
+	src: local("Countless"),
+		url(/fonts/Countless.ttf);
 }
 
 @font-face {
-  font-family: "Azonix";
-  src: local("Azonixs"),
-    url(/fonts/Azonix.otf);
+	font-family: "Azonix";
+	src: local("Azonixs"),
+		url(/fonts/Azonix.otf);
 }
 
 @font-face {
-  font-family: "Geomatrix";
-  src: local("Geomatrix"),
-    url(/fonts/GeomatrixMedium-1G4Zv.ttf);
+	font-family: "Geomatrix";
+	src: local("Geomatrix"),
+		url(/fonts/GeomatrixMedium-1G4Zv.ttf);
 }
 
 @font-face {
-  font-family: "Vensfolk";
-  src: local("Vensfolk"),
-    url(/fonts/Vensfolk.otf);
+	font-family: "Vensfolk";
+	src: local("Vensfolk"),
+		url(/fonts/Vensfolk.otf);
 }
 
 * {
-  padding: 0;
-  margin: 0;
+	padding: 0;
+	margin: 0;
 }
 
 body {
-  box-sizing: border-box;
-  font-family: "eartg";
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  background-color: #080d17;
-  font-family: 'earthorbiter';
+	box-sizing: border-box;
+	font-family: "eartg";
+	-webkit-font-smoothing: antialiased;
+	-moz-osx-font-smoothing: grayscale;
+	background-color: #080d17;
+	font-family: 'earthorbiter';
 }
 
 .giphy {
-  padding: 2px;
-  height: 70px;
-  width: 70px;
+	padding: 2px;
+	height: 70px;
+	width: 70px;
 }
 
 .Login {
-  text-align: center;
+	text-align: center;
 }
 
 .Login-logo {
-  height: 10vmin;
-  pointer-events: none;
+	height: 10vmin;
+	pointer-events: none;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-  .Login-logo {
-    animation: Login-logo-spin infinite 20s linear;
-  }
+	.Login-logo {
+		animation: Login-logo-spin infinite 20s linear;
+	}
 }
 
 .Login-header {
-  background-color: #000000;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: rgb(255, 210, 7);
+	background-color: #000000;
+	min-height: 100vh;
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	justify-content: center;
+	font-size: calc(10px + 2vmin);
+	color: rgb(255, 210, 7);
 }
 
 .Login-link {
-  color: #ff643d;
+	color: #ff643d;
 }
 
 @keyframes Login-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
+	from {
+		transform: rotate(0deg);
+	}
 
-  to {
-    transform: rotate(360deg);
-  }
+	to {
+		transform: rotate(360deg);
+	}
 }
 
 .friends {
-  color: black;
-  border-bottom: 4px;
-  border: 1px solid rgb(66, 70, 79);
-  background-color: #080d17;
-  border-radius: 0;
-  height: calc(100% - 1px);
-  overflow-y: scroll;
-  position: absolute;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  width: 15%;
+	color: black;
+	border-bottom: 4px;
+	border: 1px solid rgb(66, 70, 79);
+	background-color: #080d17;
+	border-radius: 0;
+	height: calc(100% - 1px);
+	overflow-y: scroll;
+	position: absolute;
+	top: 0;
+	left: 0;
+	bottom: 0;
+	width: 15%;
 }
 
 .typo-friends {
-  width: 100%;
-  text-align: center;
-  font-size: 1.2em;
-  margin: 10px 0;
+	width: 100%;
+	text-align: center;
+	font-size: 1.2em;
+	margin: 10px 0;
 }
 
 .yellow {
-  color: #ffc107;
+	color: #ffc107;
 }
 
 #header {
-  height: 78px;
+	height: 78px;
 }
 
 #container {
-  width: 100%;
-  height: calc(100vh - 83px);
-  display: block;
-  position: relative;
+	width: 100%;
+	height: calc(100vh - 83px);
+	display: block;
+	position: relative;
 }
 
 .launch-game {
-  background-size: cover;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  /* flex-grow: 1; */
-  position: absolute;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  width: calc(85% - 2px);
+	background-size: cover;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	/* flex-grow: 1; */
+	position: absolute;
+	right: 0;
+	top: 0;
+	bottom: 0;
+	width: calc(85% - 2px);
 }
 
 .launch-game button {
-  ;
+
+	background-image: linear-gradient(to right, #020101 0%, #740a0a  51%, #dbcc45  100%);
+	margin: 10px;
+	padding: 15px 45px;
+	text-align: center;
+	transition: 0.5s;
+	background-size: 200% auto;
+	color: #ffc107;            
+	box-shadow: 0 0 20px #080606;
+	border-radius: 10px;
+	display: block;
+	font-family: "Azonix";
+	font-size: large;
 }
+
+.launch-game button:hover {
+	background-position: right center; /* change the direction of the change here */
+	color: black;
+	text-decoration: none;
+}
+
