+++ frontend/src/components/friends/SearchBar.tsx
